sensor-kafka-project/requirements.txt
confluent-kafka==2.2.0
pandas==2.0.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
numpy==1.24.3